requirements.txt
Django==4.1.3
crispy-bootstrap5==0.7
django-jazzmin==2.6.0
django-jet-reboot==1.3.3
django-cleanup==6.0.0
django-multiselectfield==0.1.12